+++ inst/doc/plots.R
@@ -1,0 +1,8 @@
+## ----setup, include=FALSE-----------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+knitr::opts_chunk$set(echo = TRUE)
+library(reproducibleRchunks)
+