+++ inst/doc/regression_example.R
@@ -1,0 +1,10 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ----libs---------------------------------------------------------------------
+library(reproducibleRchunks)
+library(MASS)
+