+++ inst/doc/renv_example.R
@@ -1,0 +1,25 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+library(reproducibleRchunks)
+
+## ----eval=FALSE---------------------------------------------------------------
+# # Initialize renv
+# renv::init()
+
+## ----eval=FALSE---------------------------------------------------------------
+# renv::install("psych")
+# renv::install("reproducibleRchunks")
+
+## ----eval=FALSE---------------------------------------------------------------
+# renv::snapshot(prompt = FALSE)
+
+## ----eval=FALSE---------------------------------------------------------------
+# renv::restore()
+
+## ----eval=FALSE---------------------------------------------------------------
+# reproducibleRchunks::use_github_action()
+